# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CodeSentDto" : {
        "description" : "The ack body of a 'code sent' endpoint (verification request, password-reset request, contact-change request).",
        "properties" : {
          "resendAvailableAfterSeconds" : {
            "description" : "Seconds the client should wait before asking for the same code again — the server-enforced resend cooldown (0 = no cooldown configured).",
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SiteTextEntry" : {
        "properties" : {
          "key" : {
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UpdateSiteTextRequest" : {
        "properties" : {
          "texts" : {
            "items" : {
              "$ref" : "#/components/schemas/SiteTextEntry"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "VerifyRequest" : {
        "properties" : {
          "level" : {
            "enum" : [ "EMAIL", "PHONE", "SMART_ID" ],
            "type" : "string"
          }
        },
        "required" : [ "level" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "The `Authorization: Bearer <accessToken>` header read by the JWT filter. A valid token of a SUSPENDED account authenticates nothing (fresh per-request DB lookup — the DB is the truth, never a token claim).",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "aleks.bratsun@reaktiiv.com",
      "name" : "Aleks Bratsun (project owner)"
    },
    "description" : "The machine-readable contract of the OpenShelter backend. Companion to the human contracts: context-and-tasks/05-shelter-api.puml (the read contract), the README 'API' section and docs/whitepaper.md. Served at /v3/api-docs in dev/test only (SPRINGDOC_ENABLED); the committed snapshot is docs/api/openapi.json. Every non-2xx response carries the uniform ErrorResponse body.",
    "license" : {
      "name" : "MIT"
    },
    "title" : "OpenShelter API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/admin/site-texts" : {
      "put" : {
        "description" : "A batch of (key, locale) edits. A blank value resets that (key, locale) to the shipped default (the row is deleted); a blank url on a link key resets to the shipped default URL. Unknown keys, unknown locales, values over 500 characters, URLs on non-link keys and non-https URLs refuse the whole batch with a 400.",
        "operationId" : "update_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateSiteTextRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Saved"
          },
          "400" : {
            "description" : "A disallowed key/locale/value/URL"
          },
          "401" : {
            "description" : "No valid token"
          },
          "403" : {
            "description" : "Authenticated but not an admin"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "Save the admin's site-text edits",
        "tags" : [ "Admin site texts" ],
        "x-admin-only" : {
          "value" : "true"
        }
      }
    },
...
    "/api/site-texts" : {
      "get" : {
        "description" : "locale → key → {value, url?}. The three locale keys (en/et/ru) are always present (an empty map = every shipped catalog default). Absent key = the frontend catalog default. `url` exists only for the two footer source-link keys (https).",
        "operationId" : "siteTexts",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "The overrides by locale"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Validation failure — the uniform error body."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "security" : [ ],
        "summary" : "The current admin overrides for the site texts",
        "tags" : [ "Site texts" ]
      }
    },
...
    "/verify/request" : {
      "post" : {
        "description" : "202 + the resend-cooldown ack — the frontend renders a countdown instead of letting the user spam-click. SMART_ID is 400 (the provider is a stub in v1); a cooldown/cap 429 carries Retry-After in seconds.",
        "operationId" : "request",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CodeSentDto"
                }
              }
            },
            "description" : "Code sent — the ack carries the resend cooldown in seconds"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CodeSentDto"
                }
              }
            },
            "description" : "SMART_ID is not available yet (stub provider)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CodeSentDto"
                }
              }
            },
            "description" : "The level is already verified"
          },
          "429" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CodeSentDto"
                }
              }
            },
            "description" : "Cooldown / daily cap or per-IP throttle — Retry-After in seconds"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "Request a verification code",
        "tags" : [ "Account & verification" ]
      }
    }
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
}